Fixed loss of precision bug - return an int.
(from subversion r9580)

+++ greenfoot/common/SimpleTimer.java
@@ -56,6 +56,6 @@ public class SimpleTimer
      */
     public int millisElapsed()
     {
-        return System.currentTimeMillis() - lastMark;
+        return (int) (System.currentTimeMillis() - lastMark);
     }
 }
